base-defconfig: add support for ICL-Y RVP
Without some of these options the kernel does not complete the boot
process or is unstable with defconfig+base-defconfig.

Add support for Intel ME, thermal and IOMMU option. Not sure if all
options are strictly required, additional investigations could be
interesting.

The ASIX_PHY is only needed since my local RVP uses a USB dongle
instead of the Ethernet connector for some reason.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>