base-config: Enable DRM_FBDEV_EMULATION fro HDMI audio
With the 6.7.0-rc2 update preparation with the sof-kconfig generated config
we lost HDMI audio on all platforms.
It turned out that the reason for it is that DRM_FBDEV_EMULATION is no
longer enabled.
See the following commits:
bb6c4507fe82 ("drm: fix up fbdev Kconfig defaults")
a5ae331edb02b ("drm: Drop select FRAMEBUFFER_CONSOLE for DRM_FBDEV_EMULATION")
701d2054fa317 ("fbdev: Make support for userspace interfaces configurable")

and possibly more.

The connection is not clear, but we cannot afford to loose HDMI audio.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>