sof-dev-config: add CONFIG_USB_SERIAL_CP210X=m
As no one ever wanted to boot from a microphone, it's very rare that we
have to use the serial console to debug very early boot issues. But it
can happen, see for instance
https://github.com/thesofproject/linux/pull/4799

In such case it's very convenient to use another test device to collect
logs on the serial console.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>